main/requirements.txt
pandas
numpy
streamlit
streamlit_drawable_canvas
psycopg2-binary
